requirements.txt
crewai>=0.70.0
crewai-tools>=0.8.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pydantic>=2.5.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.9.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
rank-bm25>=0.2.2
anthropic>=0.39.0
gradio>=4.0.0